+++ CHANGELOG.md
@@ -9,8 +9,13 @@ feel free to ask us and community.
 
 * added support for `pg-native` for postgres (#975). To use it you just need to install `npm i pg-native` and it will be picked up automatically.
 * now Find Options support `-1` and `1` for `DESC` and `ASC` values. This is better user experience for MongoDB users. 
-* now inheritances in embeddeds are supported (#966) 
+* now inheritances in embeddeds are supported (#966).
 * `isArray: boolean` in `ColumnOptions` is deprecated. Use `array: boolean` instead.
+* deprecated `removeById` method, now use `deleteById` method instead.
+* added `insert` and `delete` methods into repository and entity manager.
+* fixed multiple issues with `update`, `updateById` and `removeById` methods in repository and entity manager. Now they do not use `save` and `remove` methods anymore - instead they are using QueryBuilder to build and execute their queries.
+* removed opencollective dependency
+* multiple small bugfixes
 
 ## 0.1.0